Create openssl.conf
The website that weather-util pulls from has some SSL oddities that seems to break TLS 1.3/1.2 (maybe?  TLS 1.2 works in the browser but not from the command line from any utility I've tried).  Later versions of Ubuntu (such as 22.04) will throw an error because "Unsafe Legacy Renegotiation" is not allowed.

This file allows us to override that.  To use this you prepend OPENSSL_CONF=/path/to/openssl.conf in front of the weather command, where openssl.conf is this file.  To note that if you stumble upon this for a different task, ensure that the data that you're using is something that you're comfortable being MITM'd as that's a potential outcome here - it's insecure.  In our case for weather it's not a huge risk so it's fine, but it's something to be aware of, at least.